Abstract correction split into separate file
