+++ probdiffeq/statespace/scalar/corr.py
@@ -5,7 +5,7 @@ import jax.numpy as jnp
 from jax.typing import ArrayLike
 
 from probdiffeq import _sqrt_util
-from probdiffeq.statespace import _collections
+from probdiffeq.statespace import _corr
 from probdiffeq.statespace.scalar import _vars
 
 
@@ -14,7 +14,7 @@ def taylor_order_zero(*args, **kwargs):
 
 
 @jax.tree_util.register_pytree_node_class
-class _TaylorZerothOrder(_collections.AbstractCorrection):
+class _TaylorZerothOrder(_corr.AbstractCorrection):
     def __repr__(self):
         return f"<TS0 with ode_order={self.ode_order}>"
 
@@ -94,7 +94,7 @@ class _TaylorZerothOrder(_collections.AbstractCorrection):
 
 
 @jax.tree_util.register_pytree_node_class
-class StatisticalFirstOrder(_collections.AbstractCorrection):
+class StatisticalFirstOrder(_corr.AbstractCorrection):
     def __init__(self, ode_order, cubature_rule):
         if ode_order > 1:
             raise ValueError
